pyproject.toml
[project]
name = "git-squared"
version = "0.1.a2"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Andrew Peters", email = "apete343@gmail.com" }
]
requires-python = ">=3.9"
dependencies = [
    "arguably>=1.3.0",
    "gitpython>=3.1.44",
    "questionary>=2.1.0",
    "rich>=13.9.4",
    "toml>=0.10.2",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "ruff>=0.9.4",
]

[project.scripts]
gg = "gg.main:main"
